pyproject.toml
[tool.poetry]
name = "bot_chat"
version = "0.1.0"
description = ""
authors = ["Konstantin Raikhert <raikhert13@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.27.1"
python-dotenv = "^0.19.2"
pre-commit = "^2.17.0"
python-telegram-bot = "^13.11"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
